requirements.txt
azure-storage-blob
python-multipart
azure-identity
scikit-learn
fastapi[all]
gunicorn
pytest
pandas
numpy
